Move as_array to field module.
This is required to be able to have a implementation of as_array for df.Field
without running into problems with circular imports.
