+++ discretisedfield/util/__init__.py
@@ -1,3 +1,3 @@
-from .util import (array2tuple, as_array, assemble_index, axesdict,
-                   bergluescher_angle, cp_hex, cp_int, fromvtk_legacy, hls2rgb,
-                   normalise_to_range, plot_box, plot_line, raxesdict)
+from .util import (array2tuple, assemble_index, axesdict, bergluescher_angle,
+                   cp_hex, cp_int, fromvtk_legacy, hls2rgb, normalise_to_range,
+                   plot_box, plot_line, raxesdict)
