+++ discretisedfield/util/util.py
@@ -1,8 +1,6 @@
 import cmath
 import collections
 import colorsys
-import functools
-import numbers
 
 import numpy as np
 
@@ -19,75 +17,6 @@ cp_int = [int(color[1:], 16) for color in cp_hex]
 
 def array2tuple(array):
     return array.item() if array.size == 1 else tuple(array.tolist())
-
-
-@functools.singledispatch
-def as_array(val, mesh, dim, dtype):
-    raise TypeError('Unsupported type {type(val)}.')
-
-
-# to avoid str being interpreted as iterable
-@as_array.register(str)
-def _(val, mesh, dim, dtype):
-    raise TypeError('Unsupported type {type(val)}.')
-
-
-@as_array.register(numbers.Complex)
-@as_array.register(collections.abc.Iterable)
-def _(val, mesh, dim, dtype):
-    if isinstance(val, numbers.Complex) and dim > 1 and val != 0:
-        raise ValueError('Wrong dimension 1 provided for value;'
-                         f' expected dimension is {dim}')
-    if dtype is None:
-        dtype = max(np.asarray(val).dtype, np.float64)
-    return np.full((*mesh.n, dim), val, dtype=dtype)
-
-
-@as_array.register(collections.abc.Callable)
-def _(val, mesh, dim, dtype):
-    # will only be called on user input
-    # dtype must be specified by the user for complex values
-    array = np.empty((*mesh.n, dim), dtype=dtype)
-    for index, point in zip(mesh.indices, mesh):
-        res = val(point)
-        array[index] = res
-    return array
-
-
-@as_array.register(dict)
-def _(val, mesh, dim, dtype):
-    # will only be called on user input
-    # dtype must be specified by the user for complex values
-    if dtype is None:
-        dtype = np.float64
-    if 'default' in val and not callable(val['default']):
-        fill_value = val['default']
-    else:
-        fill_value = np.nan
-    array = np.full((*mesh.n, dim), fill_value, dtype=dtype)
-
-    for subregion in reversed(mesh.subregions.keys()):
-        # subregions can overlap, first subregion takes precedence
-        try:
-            submesh = mesh[subregion]
-            subval = val[subregion]
-        except KeyError:
-            continue
-        else:
-            slices = mesh.region2slices(submesh.region)
-            array[slices] = as_array(subval, submesh, dim, dtype)
-
-    if np.any(np.isnan(array)):
-        # not all subregion keys specified and 'default' is missing or callable
-        if 'default' not in val:
-            raise KeyError("Key 'default' required if not all subregion keys"
-                           " are specified.")
-        subval = val['default']
-        for ix, iy, iz in np.argwhere(np.isnan(array[..., 0])):
-            # only spatial indices required -> array[..., 0]
-            array[ix, iy, iz] = subval(mesh.index2point((ix, iy, iz)))
-
-    return array
 
 
 def bergluescher_angle(v1, v2, v3):
